+++ js/main.js
@@ -3,11 +3,7 @@ import './form.js';
 import './messages.js';
 import './card.js'
 import {getData} from './request.js'
-import {renderMarkers} from './leaflet.js'
-// import {setMap} from './map.js';
-// import './data.js';
-
-// setMap();
+import {renderMarkers} from './map.js'
 
 const URL = 'https://22.javascript.pages.academy/keksobooking/data';
